Rearrange security concern
Split the code dealing with the session to a separate `SessionHandling` concern.

This way we can use the much simple `SecurityHandling` in the back office too.

Current functionality is mantained.
